requirements.txt
streamlit
numpy
pandas
scikit-learn
tensorflow
spotipy
python-dotenv
